- Fix exhaustive tests failing on client systems
The big fix here is that the make() methods in Java now throw OutOfMemoryError
as opposed to returning null.  I realized my original insistence that folks
check return values is not being followed -- even by us.  So at least for this
case we'll throw a std::bad_alloc from C++, and then translate to
java/lang/OutOfMemoryError

Also fixed the unit tests that were failing, and made sure that (a) I could
reproduce Robert's and Hudson's problems if I passed '-client' to Java, and
(b) that the edits seems to fix them.