If Michael can write code on the weekend, so can I.
FFmpeg changed their MetaData API last week and deprecated a bunch
of methods.  Now "artist", "title", "album", and "genre" are
standard tags.  Also the language attribute on the AVStream
object was removed and replace with meta-data.

Technically this change is NOT backwards compatible (for example
language designators can be longer than 4 characters) but frankly
anyone who depends on the 4 character prior limit should be taken
out back and shot anyway.

Max, I know you're watching, so this was for you.  Enjoy the
combination of C++ and Java madness!  My Excel-To-Java/C++ converter
tool went nuts.